feat: init new next app
