+++ next.config.js
@@ -1,3 +1,10 @@
-module.exports = {
-  target: 'server'
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  experimental: {
+    appDir: true,
+    forceSwcTransforms: true,
+  }
 }
+
+module.exports = nextConfig
