+++ package.json
@@ -28,6 +28,7 @@
     "badgen-icons": "^0.22.0",
     "byte-size": "^8.1.0",
     "date-fns": "^2.21.1",
+    "eslint-config-next": "^13.1.1",
     "got": "^10.7.0",
     "haxe-rpc-client": "^1.0.0",
     "humanize-duration": "^3.24.0",
@@ -37,7 +38,7 @@
     "millify": "^4.0.0",
     "my-way": "^2.0.0",
     "original-url": "^1.2.3",
-    "react-debounce-render": "^5.0.0",
+    "react-debounce-render": "^8.0.2",
     "semver": "^7.3.5",
     "serve-handler": "^6.1.3",
     "serve-marked": "^4.0.0",
@@ -47,6 +48,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.13.15",
+    "@next/font": "^13.1.1",
     "@types/fs-extra": "^9.0.11",
     "@types/lodash.debounce": "^4.0.6",
     "@types/micro": "^7.3.4",
@@ -60,7 +62,7 @@
     "@typescript-eslint/parser": "^5.16.0",
     "@zeit/next-typescript": "^1.1.1",
     "csstype": "^3.1.0",
-    "eslint": "^8.12.0",
+    "eslint": "^8.30.0",
     "eslint-plugin-react": "^7.23.2",
     "fs-extra": "^10.1.0",
     "next": "^13.1.1",
@@ -69,7 +71,7 @@
     "react-dom": "^18.2.0",
     "supertest": "^6.2.2",
     "ts-node": "^10.7.0",
-    "typescript": "^4.2.4"
+    "typescript": "^4.9.4"
   },
   "engines": {
     "node": ">=16.x",
