+++ pages/_app.tsx
@@ -1,45 +1,6 @@
-import React from 'react'
-import App from 'next/app'
-import Head from 'next/head'
+import '../styles/globals.css'
+import type { AppProps } from 'next/app'
 
-declare global {
-  interface Window {
-    dataLayer: Array<any>;
-  }
-}
-
-export default class MyApp extends App {
-  componentDidMount () {
-    window.dataLayer = window.dataLayer || []
-    function gtag (...args) { window.dataLayer.push(args) }
-    gtag('js', new Date())
-    gtag('config', 'UA-4646421-14')
-  }
-
-  render () {
-    const { Component, pageProps } = this.props
-
-    return (
-      <>
-        <Head>
-          <title>Badgen - Fast badge generating service</title>
-          <link rel='icon' type='image/png' href='/static/favicon.png' />
-          <meta name='viewport' content='initial-scale=1.0, width=device-width' />
-          <link
-            rel='stylesheet'
-            href='https://fonts.googleapis.com/css?family=Merriweather:700,300'
-          />
-          <link rel='stylesheet' href='/static/index.css' />
-          <script async src='https://www.googletagmanager.com/gtag/js?id=UA-4646421-14' />
-        </Head>
-        <Component {...pageProps} />
-        <style jsx global>{`
-          html, body { margin: 0; height: 100%; scroll-behavior: smooth }
-          #__next { height: 100% }
-          a { text-decoration: none }
-        `}
-        </style>
-      </>
-    )
-  }
+export default function App({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />
 }
